OPSMN-5282: Associated the temporal expression to the right top-level form for the wide row calculation.
When a temporal expression involves fields of 1:many relation then the expression
should be associated or attached to the many side of the relation so that
all values of expression for a given parent record appear in the output
row.

For example - when a temporal expression involves a field of participant and
visit, the expression should be associated with the visit level form so that
all expression values, calculated for each visit of the participant appear in
the output row. If it is incorrectly associated with the participant form,
as was done before this fix, then only one expression value appears for all
the participant visits.